+++ Assets/Scripts/CardData/UnicornCardDatas/FreeCandyUnicornCardData.cs
@@ -1,0 +1,18 @@
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(menuName = "CardData/UnicornCardData/FreeCandyUnicornCardData")]
+public class FreeCandyUnicornCardData : UnicornCardData
+{
+    public override void OnEnable()
+    {
+        base.OnEnable();
+        unicornType = UnicornType.MAGIC;
+        specialActionType = SpecialActionType.IMMEDIATE;
+
+        actions = new List<CardAction>
+        {
+            new StealUnicornAction { targetSubtype = UnicornType.BABY }
+        };
+    }
+}